encari: Support dense classification outputs
As a better alternative to inplace_relabel, it is now possible to return
a napari Label layer with the relabeled segmentation + classification
map instead of returning a Shapes layer or nothing.

The new default shape type "dense" enables this feature. Actual shapes
(rectangles, ellipses) can still be chosen in the dropdown.